backend/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
joblib==1.0.1
numpy==1.19.5
pandas==1.1.4
statsmodels==0.12.1
scikit-learn==0.24.1
django-cors-headers==3.7.0
mlflow
google-cloud
gs
google-cloud-storage
tensorflow